change names in circular dependency tests
