+++ lib/src/apiTest/kotlin/com/lemonappdev/konsist/architecture/dependencyrules/circulardependency1/CircularDependency1Test.kt
@@ -1,4 +1,4 @@
-package com.lemonappdev.konsist.architecture.dependencyrules.circulardependency5
+package com.lemonappdev.konsist.architecture.dependencyrules.circulardependency1
 
 import com.lemonappdev.konsist.api.architecture.KoArchitectureCreator.architecture
 import com.lemonappdev.konsist.api.architecture.Layer
@@ -7,32 +7,30 @@ import org.amshove.kluent.shouldThrow
 import org.amshove.kluent.withMessage
 import org.junit.jupiter.api.Test
 
-class CircularDependency5Test {
+class CircularDependency1Test {
     @Test
-    fun `circular dependency 5`() {
+    fun `circular dependency 1`() {
         // given
         val layer1 = Layer("layer1", "layer1..")
         val layer2 = Layer("layer2", "layer2..")
         val layer3 = Layer("layer3", "layer3..")
-        val layer4 = Layer("layer4", "layer4..")
 
         // when
         val sut = {
             architecture {
-                layer1.dependsOn(layer2, layer3)
-                layer2.dependsOn(layer4)
-                layer3.dependsOn(layer4)
-                layer4.dependsOn(layer1)
+                layer1.dependsOn(layer2)
+                layer2.dependsOn(layer3)
+                layer3.dependsOn(layer1)
             }
         }
 
         // then
         sut shouldThrow KoPreconditionFailedException::class withMessage """
             Illegal circular dependencies:
-            Layer layer4 -->
+            Layer layer3 -->
             Layer layer1 -->
             Layer layer2 -->
-            Layer layer4.
+            Layer layer3.
         """.trimIndent()
     }
 }
