+++ lib/src/apiTest/kotlin/com/lemonappdev/konsist/architecture/dependencyrules/circulardependency3/CircularDependency3Test.kt
@@ -15,25 +15,23 @@ class CircularDependency3Test {
         val layer2 = Layer("layer2", "layer2..")
         val layer3 = Layer("layer3", "layer3..")
         val layer4 = Layer("layer4", "layer4..")
-        val layer5 = Layer("layer5", "layer5..")
 
         // when
         val sut = {
             architecture {
-                layer1.dependsOn(layer2, layer3)
-                layer2.dependsOn(layer4, layer5)
-                layer3.dependsOn(layer5)
-                layer4.dependsOn(layer1)
+                layer1.dependsOn(layer2)
+                layer2.dependsOn(layer3)
+                layer3.dependsOn(layer1, layer4)
             }
         }
 
         // then
         sut shouldThrow KoPreconditionFailedException::class withMessage """
             Illegal circular dependencies:
-            Layer layer4 -->
+            Layer layer3 -->
             Layer layer1 -->
             Layer layer2 -->
-            Layer layer4.
+            Layer layer3.
         """.trimIndent()
     }
 }
