+++ lib/src/apiTest/kotlin/com/lemonappdev/konsist/architecture/dependencyrules/circulardependency4/README.md
@@ -5,7 +5,8 @@ Test scenario with circular dependency:
 ```mermaid
 %%{init: {'theme': 'forest'}}%%
 flowchart LR
-    Layer1 --> Layer2 
-    Layer2 --> Layer3
-    Layer3 --> Layer1 & Layer4
+    Layer1 --> Layer2 & Layer3
+    Layer2 --> Layer4
+    Layer3 --> Layer4
+    Layer4 --> Layer1
 ```
