--- lib/src/apiTest/kotlin/com/lemonappdev/konsist/architecture/dependencyrules/circulardependency5/README.md
@@ -1,12 +1,0 @@
-# Circular dependency 5
-
-Test scenario with circular dependency:
-
-```mermaid
-%%{init: {'theme': 'forest'}}%%
-flowchart LR
-    Layer1 --> Layer2 & Layer3
-    Layer2 --> Layer4
-    Layer3 --> Layer4
-    Layer4 --> Layer1
-```
